Change algorythm used for bisection
After a lot of testing we changed the approach from looking for the
first good to looking for the first bad. This usually represents the
regression finding better and converges faster.

We also added check whether individual axis are making a difference.
This speed-ups finding the issues rapidly, also it might cover some
dependencies that require different setting of the later axis, which is
aligned with the aim of this project.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>